Refactor Authorize component.
